added query term test and fixed delete folder with $unset

+++ routes/notes.js
@@ -6,6 +6,7 @@ const router = express.Router();
 
 const Note = require('../models/note');
 
+const mongoose = require('mongoose');
 
 
 
@@ -13,6 +14,8 @@ const Note = require('../models/note');
 router.get('/', (req, res, next) => {
   // console.log("proccess env var: ", process.env);
   const searchTerm = req.query.searchTerm;
+  const folderId = req.query.folderId;
+  // console.log('folderId is', folderId);
   // console.log(searchTerm);
   let filter = {};
   let query = {};
@@ -20,13 +23,20 @@ router.get('/', (req, res, next) => {
   if (searchTerm) {
     filter.title = re,
     filter.content = re;
-    query = {$or : [{title: filter.title}, {content: filter.content}]};
+    query = {$or : [{title: filter.title},{content: filter.content}]};
   }
+  if (folderId){
+
+    filter.folderId = folderId;
+    Object.keys(query).length === 0 ? query = {folderId: filter.folderId} : query.$or.push({folderId: filter.folderId});
+  }
+  
   Note
     .find(query)
     .sort({ updatedAt: 'desc' })  
     .then((results) => {
       res.json(results);
+      console.log('folder is: ',results[0].folderId);
     })
     .catch(err => {
       next(err);
@@ -39,6 +49,12 @@ router.get('/', (req, res, next) => {
 router.get('/:id', (req, res, next) => {
   const id = req.params.id;
 
+  if(!mongoose.Types.ObjectId.isValid(req.params.id)){
+    const err = new Error('please enter valid id!');
+    err.status = 400;
+    return next(err);
+  }
+
   Note.findById(id)
     .then((result) => {
       res.json(result);
@@ -47,17 +63,31 @@ router.get('/:id', (req, res, next) => {
       next(err);
     });
 });
+// function validateFolder(id){
+//   if (!mongoose.Types.ObjectId.isValid(folderId)){
+//     const err = new Error ('Please input valid folder id');
+//     err.status = (400);
+//     return err;
+//   }else return true;
+// }
 
 /* ========== POST/CREATE AN ITEM ========== */
 router.post('/', (req, res, next) => {
-  const {title, content} = req.body;
+  const {title, content, folderId} = req.body;
   //validate user input
   if(!title){
     const err = new Error('title is required!');
     err.status = (400);
     next(err);
   }
-  const newNote = {title, content};
+  if(folderId){
+    if (!mongoose.Types.ObjectId.isValid(folderId)){
+      const err = new Error ('Please input valid folder id');
+      err.status = (400);
+      return next(err);
+    }
+  }
+  const newNote = {title, content, folderId};
 
   Note
     .create(newNote)
@@ -72,19 +102,38 @@ router.post('/', (req, res, next) => {
 /* ========== PUT/UPDATE A SINGLE ITEM ========== */
 router.put('/:id', (req, res, next) => {
   const id = req.params.id;
-  const {title, content} = req.body;
-  const updateNote = {title, content};
-
+  const {title, content, folderId} = req.body;
+  const updateNote = {title, content, folderId};
+  
+  if(folderId){
+    if(!mongoose.Types.ObjectId.isValid(req.params.id)){
+      const err = new Error('please enter valid id!');
+      err.status = 400;
+      return next(err);
+    }
+  }
+  if(!updateNote.folderId){
+    delete updateNote.folderId;
+    updateNote.$unset = {folderId : 1};
+  }
   Note
-    .findByIdAndUpdate(id, updateNote, {new: true})
+    .findByIdAndUpdate(id, updateNote, {new: true})  
     .then((result) =>  {
-      res.location(`api/notes/${result.id}`).status(201).json(result);
+      res.location(`api/notes/${result.id}`).status(202).json(result);
     })
     .catch(err => next(err));
+  
+
 });
 
 /* ========== DELETE/REMOVE A SINGLE ITEM ========== */
 router.delete('/:id', (req, res, next) => {
+
+  if(!mongoose.Types.ObjectId.isValid(req.params.id)){
+    const err = new Error('please enter valid id!');
+    err.status = 400;
+    return next(err);
+  }
   const id = req.params.id;
   Note.findByIdAndRemove(id)
     .then(() => res.status(204).end())
